backend/foodgram/api/mixins.py
from rest_framework import viewsets


class ExcludeUpdateModelViewSet(viewsets.ModelViewSet):
    http_method_names = ['get', 'post', 'patch', 'delete']

    def perform_create(self, serializer):
        serializer.save(author=self.request.user)
